refactor(core): split hook executor boundaries
