+++ crates/git-smee-core/src/executor.rs
@@ -1,26 +1,15 @@
-use std::{
-    env,
-    io::{self, ErrorKind, Write},
-    process::Stdio,
-    sync::Mutex,
-    thread,
-    time::{Duration, Instant},
-};
-
-#[cfg(windows)]
-use std::os::windows::process::CommandExt;
-#[cfg(windows)]
-use std::path::PathBuf;
-
-use rayon::iter::IntoParallelRefIterator;
-use rayon::prelude::*;
 use thiserror::Error;
 
-use crate::{
-    SmeeConfig,
-    config::{HookDefinition, LifeCyclePhase},
-    platform::Platform,
-};
+mod redaction;
+mod runner;
+mod scheduler;
+mod summary;
+
+use crate::{SmeeConfig, config::LifeCyclePhase, platform::Platform};
+
+use runner::{CommandRunner, PlatformCommandRunner};
+use scheduler::{run_hooks_with_runner, run_hooks_with_runner_with_summary};
+pub use summary::{CommandPhase, HookRunSummary};
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -141,599 +130,16 @@ fn execute_hook_with_runner_and_summary<R: CommandRunner>(
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CommandPhase {
-    Sequential,
-    Parallel,
-}
-
-impl CommandPhase {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Sequential => "sequential",
-            Self::Parallel => "parallel",
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct HookRunSummary {
-    total_configured: usize,
-    total_duration: Duration,
-    sequential_duration: Duration,
-    parallel_duration: Duration,
-    command_runs: Vec<CommandRun>,
-}
-
-impl HookRunSummary {
-    pub fn total_configured(&self) -> usize {
-        self.total_configured
-    }
-
-    pub fn attempted_count(&self) -> usize {
-        self.command_runs.len()
-    }
-
-    pub fn skipped_count(&self) -> usize {
-        self.total_configured.saturating_sub(self.attempted_count())
-    }
-
-    pub fn failed_count(&self) -> usize {
-        self.command_runs
-            .iter()
-            .filter(|run| run.outcome.is_failure())
-            .count()
-    }
-
-    pub fn phase_attempted_count(&self, phase: CommandPhase) -> usize {
-        self.command_runs
-            .iter()
-            .filter(|run| run.phase == phase)
-            .count()
-    }
-
-    pub fn phase_failed_count(&self, phase: CommandPhase) -> usize {
-        self.command_runs
-            .iter()
-            .filter(|run| run.phase == phase && run.outcome.is_failure())
-            .count()
-    }
-
-    pub fn first_failure(&self) -> Option<&CommandRun> {
-        self.command_runs
-            .iter()
-            .filter(|run| run.outcome.is_failure())
-            .min_by_key(|run| (run.phase_sort_key(), run.index))
-    }
-
-    pub fn error(&self) -> Option<Error> {
-        self.first_failure().and_then(CommandRun::to_error)
-    }
-
-    pub fn text_lines(&self, phase: LifeCyclePhase) -> Vec<String> {
-        let mut lines = vec![
-            format!("Hook summary: {phase}"),
-            format!(
-                "  total: {} attempted, {} skipped, {} failed in {}",
-                self.attempted_count(),
-                self.skipped_count(),
-                self.failed_count(),
-                format_duration(self.total_duration),
-            ),
-            format!(
-                "  sequential: {} attempted, {} failed in {}",
-                self.phase_attempted_count(CommandPhase::Sequential),
-                self.phase_failed_count(CommandPhase::Sequential),
-                format_duration(self.sequential_duration),
-            ),
-            format!(
-                "  parallel: {} attempted, {} failed in {}",
-                self.phase_attempted_count(CommandPhase::Parallel),
-                self.phase_failed_count(CommandPhase::Parallel),
-                format_duration(self.parallel_duration),
-            ),
-        ];
-        for run in &self.command_runs {
-            lines.push(format!(
-                "  - {} command #{}: {} in {}",
-                run.phase.as_str(),
-                run.index + 1,
-                run.status_display(),
-                format_duration(run.duration),
-            ));
-        }
-        if let Some(first_failure) = self.first_failure() {
-            lines.push(format!(
-                "  first failure: {}",
-                first_failure.failure_display()
-            ));
-        }
-        lines
-    }
-}
-
-#[derive(Debug)]
-pub struct CommandRun {
-    phase: CommandPhase,
-    index: usize,
-    duration: Duration,
-    outcome: CommandOutcome,
-}
-
-impl CommandRun {
-    const fn phase_sort_key(&self) -> usize {
-        match self.phase {
-            CommandPhase::Sequential => 0,
-            CommandPhase::Parallel => 1,
-        }
-    }
-
-    fn status_display(&self) -> String {
-        match &self.outcome {
-            CommandOutcome::Success => "ok".to_string(),
-            CommandOutcome::Exit(code) => format!("failed with code {code}"),
-            CommandOutcome::Signal => "terminated by signal".to_string(),
-            CommandOutcome::SpawnFailed { .. } => "spawn failed".to_string(),
-            CommandOutcome::NoCommandDefined => "no command defined".to_string(),
-        }
-    }
-
-    fn failure_display(&self) -> String {
-        let prefix = format!("{} command #{}", self.phase.as_str(), self.index + 1);
-        match &self.outcome {
-            CommandOutcome::Success => format!("{prefix} succeeded"),
-            CommandOutcome::Exit(code) => format!("{prefix} exited with code {code}"),
-            CommandOutcome::Signal => format!("{prefix} was terminated by a signal"),
-            CommandOutcome::SpawnFailed {
-                command,
-                shell,
-                source,
-            } => {
-                format!("{prefix} failed to spawn '{command}' via '{shell}': {source}")
-            }
-            CommandOutcome::NoCommandDefined => format!("{prefix} had no command defined"),
-        }
-    }
-
-    fn to_error(&self) -> Option<Error> {
-        match &self.outcome {
-            CommandOutcome::Success => None,
-            CommandOutcome::Exit(code) => Some(Error::ExecutionFailed(*code)),
-            CommandOutcome::Signal => Some(Error::ExecutionTerminatedBySignal),
-            CommandOutcome::SpawnFailed {
-                command,
-                shell,
-                source,
-            } => Some(Error::CommandSpawnFailed {
-                command: command.clone(),
-                shell: shell.clone(),
-                source: io::Error::new(source.kind(), source.to_string()),
-            }),
-            CommandOutcome::NoCommandDefined => Some(Error::NoCommandDefined),
-        }
-    }
-}
-
-#[derive(Debug)]
-enum CommandOutcome {
-    Success,
-    Exit(i32),
-    Signal,
-    SpawnFailed {
-        command: String,
-        shell: String,
-        source: io::Error,
-    },
-    NoCommandDefined,
-}
-
-impl CommandOutcome {
-    const fn is_failure(&self) -> bool {
-        !matches!(self, Self::Success)
-    }
-}
-
-trait CommandRunner: Sync {
-    fn run(
-        &self,
-        command: &str,
-        hook_args: &[String],
-        stdin_payload: Option<&[u8]>,
-    ) -> Result<Option<i32>, std::io::Error>;
-    fn shell_display(&self) -> &'static str;
-}
-
-struct PlatformCommandRunner<'a> {
-    platform: &'a Platform,
-}
-
-impl CommandRunner for PlatformCommandRunner<'_> {
-    fn run(
-        &self,
-        command: &str,
-        hook_args: &[String],
-        stdin_payload: Option<&[u8]>,
-    ) -> Result<Option<i32>, std::io::Error> {
-        let mut shell_command = self.platform.create_command();
-        apply_hook_arg_env(&mut shell_command, hook_args);
-        let mut _windows_command_script = None;
-        match self.platform {
-            Platform::Unix => {
-                shell_command.arg(command);
-                shell_command.arg("--");
-                shell_command.args(hook_args);
-            }
-            Platform::Windows => {
-                let command_script = create_windows_command_script(command)?;
-                shell_command.arg(&command_script);
-                #[cfg(windows)]
-                append_windows_hook_args(&mut shell_command, hook_args);
-                #[cfg(not(windows))]
-                shell_command.args(hook_args);
-                _windows_command_script = Some(command_script);
-            }
-        }
-        if stdin_payload.is_some() {
-            shell_command.stdin(Stdio::piped());
-        }
-
-        #[cfg(windows)]
-        if let Some(current_dir) = cmd_compatible_current_dir()? {
-            shell_command.current_dir(current_dir);
-        }
-
-        let mut child = shell_command.spawn()?;
-        if let Some(stdin_payload) = stdin_payload {
-            let Some(mut stdin) = child.stdin.take() else {
-                return child.wait().map(|status| status.code());
-            };
-            let stdin_payload = stdin_payload.to_vec();
-            let stdin_writer = thread::spawn(move || {
-                match stdin.write_all(&stdin_payload) {
-                    // Hook commands are allowed to ignore or close stdin early. If the
-                    // command exits successfully, a broken pipe while replaying the
-                    // buffered hook payload should not fail the hook run.
-                    Err(error) if error.kind() == ErrorKind::BrokenPipe => Ok(()),
-                    result => result,
-                }
-            });
-            let wait_result = child.wait().map(|status| status.code());
-            let stdin_result = stdin_writer
-                .join()
-                .unwrap_or_else(|_| Err(io::Error::other("stdin writer thread panicked")));
-            stdin_result?;
-            wait_result
-        } else {
-            child.wait().map(|status| status.code())
-        }
-    }
-
-    fn shell_display(&self) -> &'static str {
-        self.platform.shell_display()
-    }
-}
-
-fn create_windows_command_script(command: &str) -> Result<tempfile::TempPath, io::Error> {
-    let mut script = tempfile::Builder::new()
-        .prefix("git-smee-command-")
-        .suffix(".cmd")
-        .tempfile()?;
-    script.write_all(windows_command_script(command).as_bytes())?;
-    script.flush()?;
-    Ok(script.into_temp_path())
-}
-
-#[cfg(windows)]
-fn append_windows_hook_args(shell_command: &mut std::process::Command, hook_args: &[String]) {
-    for arg in hook_args {
-        shell_command.raw_arg(" ");
-        shell_command.raw_arg(windows_cmd_quote_hook_arg(arg));
-    }
-}
-
-#[cfg_attr(not(windows), allow(dead_code))]
-fn windows_cmd_quote_hook_arg(arg: &str) -> String {
-    let needs_quotes = arg.is_empty()
-        || arg.chars().any(|ch| {
-            matches!(
-                ch,
-                ' ' | '\t' | '&' | '|' | '^' | '<' | '>' | '(' | ')' | '!' | '%'
-            )
-        });
-    if !needs_quotes {
-        return arg.to_string();
-    }
-
-    let escaped = arg.replace('"', "\"\"");
-    format!("\"{escaped}\"")
-}
-
-fn windows_command_script(command: &str) -> String {
-    format!("@echo off\r\n{command}\r\n")
-}
-
-fn apply_hook_arg_env(shell_command: &mut std::process::Command, hook_args: &[String]) {
-    for (key, _) in env::vars_os() {
-        if is_hook_arg_env_key(&key.to_string_lossy()) {
-            #[cfg(windows)]
-            shell_command.env_remove(key.to_string_lossy().to_ascii_uppercase());
-            #[cfg(not(windows))]
-            shell_command.env_remove(key);
-        }
-    }
-
-    shell_command.env("GIT_SMEE_HOOK_ARGC", hook_args.len().to_string());
-    for (index, arg) in hook_args.iter().enumerate() {
-        shell_command.env(format!("GIT_SMEE_HOOK_ARG_{}", index + 1), arg);
-    }
-}
-
-#[cfg(windows)]
-fn cmd_compatible_current_dir() -> Result<Option<PathBuf>, std::io::Error> {
-    let current_dir = env::current_dir()?;
-    let current_dir = current_dir.to_string_lossy();
-    Ok(current_dir.strip_prefix(r"\\?\").map(PathBuf::from))
-}
-
-fn is_hook_arg_env_key(key: &str) -> bool {
-    if key.eq_ignore_ascii_case("GIT_SMEE_HOOK_ARGC") {
-        return true;
-    }
-
-    let prefix = "GIT_SMEE_HOOK_ARG_";
-    let Some(prefix_part) = key.get(..prefix.len()) else {
-        return false;
-    };
-    if !prefix_part.eq_ignore_ascii_case(prefix) {
-        return false;
-    }
-    let suffix = &key[prefix.len()..];
-    !suffix.is_empty() && !suffix.starts_with('0') && suffix.chars().all(|ch| ch.is_ascii_digit())
-}
-
-type IndexedHook<'a> = (usize, &'a HookDefinition);
-
-fn run_hooks_with_runner<R: CommandRunner>(
-    hooks: &[HookDefinition],
-    runner: &R,
-    hook_args: &[String],
-    stdin_payload: Option<&[u8]>,
-) -> Result<(), Error> {
-    let summary = run_hooks_with_runner_with_summary(hooks, runner, hook_args, stdin_payload);
-    match summary.error() {
-        Some(error) => Err(error),
-        None => Ok(()),
-    }
-}
-
-fn run_hooks_with_runner_with_summary<R: CommandRunner>(
-    hooks: &[HookDefinition],
-    runner: &R,
-    hook_args: &[String],
-    stdin_payload: Option<&[u8]>,
-) -> HookRunSummary {
-    let started = Instant::now();
-    let (parallel_hooks, sequential_hooks): (Vec<IndexedHook<'_>>, Vec<IndexedHook<'_>>) = hooks
-        .iter()
-        .enumerate()
-        .partition(|(_, hook)| hook.parallel_execution_allowed);
-
-    let mut command_runs = Vec::new();
-    let mut failed = false;
-    let sequential_started = Instant::now();
-    for (phase_index, (_, hook)) in sequential_hooks.into_iter().enumerate() {
-        if failed {
-            break;
-        }
-        let run = execute_command_record(
-            CommandPhase::Sequential,
-            phase_index,
-            &hook.command,
-            runner,
-            hook_args,
-            stdin_payload,
-        );
-        failed = run.outcome.is_failure();
-        command_runs.push(run);
-    }
-    let sequential_duration = sequential_started.elapsed();
-
-    let mut parallel_duration = Duration::ZERO;
-    if !failed {
-        let parallel_started = Instant::now();
-        let parallel_runs = Mutex::new(Vec::new());
-        let _ = parallel_hooks
-            .par_iter()
-            .enumerate()
-            .try_for_each(|(phase_index, (_, hook))| {
-                let run = execute_command_record(
-                    CommandPhase::Parallel,
-                    phase_index,
-                    &hook.command,
-                    runner,
-                    hook_args,
-                    stdin_payload,
-                );
-                let failed = run.outcome.is_failure();
-                lock_command_runs(&parallel_runs).push(run);
-                if failed { Err(()) } else { Ok(()) }
-            });
-        let mut parallel_runs = match parallel_runs.into_inner() {
-            Ok(runs) => runs,
-            Err(poisoned) => poisoned.into_inner(),
-        };
-        parallel_duration = parallel_started.elapsed();
-        parallel_runs.sort_by_key(|run| run.index);
-        command_runs.extend(parallel_runs);
-    }
-
-    HookRunSummary {
-        total_configured: hooks.len(),
-        total_duration: started.elapsed(),
-        sequential_duration,
-        parallel_duration,
-        command_runs,
-    }
-}
-
-fn lock_command_runs(
-    command_runs: &Mutex<Vec<CommandRun>>,
-) -> std::sync::MutexGuard<'_, Vec<CommandRun>> {
-    match command_runs.lock() {
-        Ok(guard) => guard,
-        Err(poisoned) => poisoned.into_inner(),
-    }
-}
-
-fn execute_command_record(
-    phase: CommandPhase,
-    index: usize,
-    command: &str,
-    runner: &impl CommandRunner,
-    hook_args: &[String],
-    stdin_payload: Option<&[u8]>,
-) -> CommandRun {
-    let started = Instant::now();
-    let outcome = if command.trim().is_empty() {
-        CommandOutcome::NoCommandDefined
-    } else {
-        match runner.run(command, hook_args, stdin_payload) {
-            Ok(Some(0)) => CommandOutcome::Success,
-            Ok(Some(exit_status_code)) => CommandOutcome::Exit(exit_status_code),
-            Ok(None) => CommandOutcome::Signal,
-            Err(source) => CommandOutcome::SpawnFailed {
-                command: redact_command(command),
-                shell: runner.shell_display().to_string(),
-                source,
-            },
-        }
-    };
-    CommandRun {
-        phase,
-        index,
-        duration: started.elapsed(),
-        outcome,
-    }
-}
-
-fn format_duration(duration: Duration) -> String {
-    if duration.as_secs() > 0 {
-        format!("{:.2}s", duration.as_secs_f64())
-    } else {
-        format!("{}ms", duration.as_millis())
-    }
-}
-
-#[cfg(test)]
-fn execute_command(
-    command: &str,
-    runner: &impl CommandRunner,
-    hook_args: &[String],
-    stdin_payload: Option<&[u8]>,
-) -> Result<(), Error> {
-    if command.trim().is_empty() {
-        return Err(Error::NoCommandDefined);
-    }
-    let exit_code = runner
-        .run(command, hook_args, stdin_payload)
-        .map_err(|source| Error::CommandSpawnFailed {
-            command: redact_command(command),
-            shell: runner.shell_display().to_string(),
-            source,
-        })?;
-    match exit_code {
-        Some(0) => Ok(()),
-        Some(exit_status_code) => Err(Error::ExecutionFailed(exit_status_code)),
-        None => Err(Error::ExecutionTerminatedBySignal),
-    }
-}
-
-fn redact_command(command: &str) -> String {
-    let tokens = tokenize_command(command);
-    let executable_index = tokens
-        .iter()
-        .position(|token| !is_inline_env_assignment(token));
-    let mut redacted = executable_index
-        .and_then(|index| tokens.get(index))
-        .cloned()
-        .unwrap_or_else(|| "<redacted>".to_string());
-    if redacted.chars().count() > 80 {
-        redacted = redacted.chars().take(77).collect();
-        redacted.push_str("...");
-    }
-    if let Some(index) = executable_index
-        && tokens.len() > index + 1
-    {
-        redacted.push_str(" <args redacted>");
-    }
-    redacted
-}
-
-fn is_inline_env_assignment(token: &str) -> bool {
-    let Some((key, _)) = token.split_once('=') else {
-        return false;
-    };
-    is_valid_env_var_name(key)
-}
-
-fn is_valid_env_var_name(value: &str) -> bool {
-    let mut chars = value.chars();
-    let Some(first) = chars.next() else {
-        return false;
-    };
-    if !matches!(first, 'A'..='Z' | 'a'..='z' | '_') {
-        return false;
-    }
-    chars.all(|ch| matches!(ch, 'A'..='Z' | 'a'..='z' | '0'..='9' | '_'))
-}
-
-fn tokenize_command(command: &str) -> Vec<String> {
-    let mut tokens = Vec::new();
-    let mut current = String::new();
-    let mut in_single_quotes = false;
-    let mut in_double_quotes = false;
-    let mut chars = command.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        match ch {
-            '\\' if !in_single_quotes => {
-                current.push(ch);
-                if let Some(next) = chars.peek().copied()
-                    && (next.is_whitespace() || matches!(next, '\\' | '\'' | '"'))
-                {
-                    current.push(chars.next().expect("peeked char should exist"));
-                }
-            }
-            '\'' if !in_double_quotes => {
-                in_single_quotes = !in_single_quotes;
-            }
-            '"' if !in_single_quotes => {
-                in_double_quotes = !in_double_quotes;
-            }
-            ch if ch.is_whitespace() && !in_single_quotes && !in_double_quotes => {
-                if !current.is_empty() {
-                    tokens.push(std::mem::take(&mut current));
-                }
-            }
-            _ => current.push(ch),
-        }
-    }
-
-    if !current.is_empty() {
-        tokens.push(current);
-    }
-
-    tokens
-}
-
 #[cfg(test)]
 mod tests {
     use std::{
         collections::{HashMap, VecDeque},
+        env,
         ffi::OsString,
         io,
         process::Command,
         sync::{Arc, Barrier, Mutex},
+        time::Duration,
     };
 
     use assert2::assert;
@@ -741,6 +147,12 @@ mod tests {
 
     use crate::{config::HookDefinition, test_support::process_state_lock};
 
+    use super::redaction::redact_command;
+    use super::runner::{
+        apply_hook_arg_env, is_hook_arg_env_key, windows_cmd_quote_hook_arg, windows_command_script,
+    };
+    use super::scheduler::execute_command;
+    use super::summary::{CommandOutcome, CommandRun};
     use super::*;
 
     enum PlannedResult {
